fix: types, edit, components
