+++ frontend/src/data/mock-data.ts
@@ -1,18 +1,4 @@
-export type Variation = {
-  key: string;
-  value: boolean | number | string;
-};
-
-export type FeatureFlag = {
-  id: string;
-  key: string;
-  createdAt: string;
-  createdBy: string;
-  enabled: boolean;
-  type: 'Bool' | 'Number' | 'String';
-  variations: Variation[];
-  defaultValue: boolean | number | string;
-};
+import { FeatureFlag } from '@/types/featureFlag';
 
 export const mockFeatureFlags: FeatureFlag[] = [
   {
